server/requirements.txt
alembic==1.18.0
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
Flask==3.1.2
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
SQLAlchemy==2.0.45
tomli==2.4.0
typing_extensions==4.15.0
Werkzeug==3.1.5
